Routing, Navbar, Advanced Component
work time: 2h
